Replace `unsigned long long` with `uint64_t`
`uint64_t` is already being used everywhere and I much prefer the
explicitly sized types.

+++ src/pevents.h
@@ -23,7 +23,7 @@ namespace neosmart {
     typedef neosmart_event_t_ *neosmart_event_t;
 
     // Constant declarations
-    const uint64_t WAIT_INFINITE = ~((unsigned long long)0);
+    const uint64_t WAIT_INFINITE = ~((uint64_t)0);
 
     // Function declarations
     neosmart_event_t CreateEvent(bool manualReset = false, bool initialState = false);
